Added :camera: option
With the :camera: option it is possible to set the position of the camera for 3d plots.
This allows to create 3D plots with the intented view (otherwise, useful features might be hidden
by other objects).

The :camera: options accepts 9 float arguments:

* x_cam, y_cam, z_cam: the position of the camera.
* x_tar, y_tar, z_tar: the position of the target (where the camera is looking at).
* x_up, y_up, z_up: the up direction.

Currently, it only works for code blocks where no fig.show() is used and when fig-vars is
not set.